# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: City Stories Guide API
  description: 'REST API for City Stories Guide — an ambient audio guide app that tells stories

    about nearby places while walking. Provides endpoints for discovering nearby stories,

    managing cities/POIs/stories, authentication, purchases, and admin operations.

    '
  version: 1.0.0
  contact:
    name: City Stories Guide Team
servers:
- url: /api/v1
  description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from auth endpoints
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          example: invalid request
        trace_id:
          type: string
          description: Request trace ID for debugging
          example: abc-123-def
    ValidationDetail:
      type: object
      required:
      - field
      - message
      properties:
        field:
          type: string
          description: JSON field name that failed validation
          example: email
        message:
          type: string
          description: Human-readable validation failure reason
          example: must be a valid email address
    ValidationError:
      type: object
      required:
      - error
      - details
      properties:
        error:
          type: string
          enum:
          - validation_error
          example: validation_error
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationDetail'
        trace_id:
          type: string
          description: Request trace ID for debugging
          example: abc-123-def
    ErrorOrValidationError:
      oneOf:
      - $ref: '#/components/schemas/Error'
      - $ref: '#/components/schemas/ValidationError'
...
    ReportResponse:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/Report'
...
    PurchaseResponse:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/Purchase'
...
    CreateReportRequest:
      type: object
      required:
      - story_id
      - user_id
      - type
      properties:
        story_id:
          type: integer
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - wrong_location
          - wrong_fact
          - inappropriate_content
        comment:
          type: string
          minLength: 10
          maxLength: 1000
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
        lng:
          type: number
          format: double
          minimum: -180
          maximum: 180
...
    VerifyPurchaseRequest:
      type: object
      required:
      - platform
      - transaction_id
      - receipt
      - type
      - price
      properties:
        platform:
          type: string
          enum:
          - ios
          - android
          example: ios
        transaction_id:
          type: string
          maxLength: 500
          example: txn_abc123
        receipt:
          type: string
          maxLength: 50000
        type:
          type: string
          enum:
          - city_pack
          - subscription
          - lifetime
        city_id:
          type: integer
          nullable: true
          description: Optional city ID for city_pack purchases.
        price:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          example: 4.99
...
    Report:
      type: object
      required:
      - id
      - story_id
      - user_id
      - type
      - status
      - created_at
      properties:
        id:
          type: integer
          example: 1
        story_id:
          type: integer
          example: 5
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - wrong_location
          - wrong_fact
          - inappropriate_content
          example: wrong_fact
        comment:
          type: string
          nullable: true
          example: The date mentioned is incorrect
        user_lat:
          type: number
          format: double
          nullable: true
        user_lng:
          type: number
          format: double
          nullable: true
        status:
          type: string
          enum:
          - new
          - reviewed
          - resolved
          - dismissed
          example: new
        resolved_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Purchase:
      type: object
      required:
      - id
      - user_id
      - type
      - platform
      - price
      - is_ltd
      - created_at
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - city_pack
          - subscription
          - lifetime
          example: city_pack
        city_id:
          type: integer
          nullable: true
          example: 1
        platform:
          type: string
          enum:
          - ios
          - android
          example: ios
        transaction_id:
          type: string
          nullable: true
          example: txn_abc123
        price:
          type: number
          format: double
          example: 4.99
        is_ltd:
          type: boolean
          example: false
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /reports:
    post:
      tags:
      - Reports
      summary: Create a content report
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
            example:
              story_id: 5
              user_id: "550e8400-e29b-41d4-a716-446655440000"
              type: "wrong_fact"
              comment: "The date mentioned in the story is incorrect — the fortress was built in the 4th century, not the 6th."
              lat: 41.6875
              lng: 44.8084
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
              example:
                data:
                  id: 1
                  story_id: 5
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  type: "wrong_fact"
                  comment: "The date mentioned in the story is incorrect — the fortress was built in the 4th century, not the 6th."
                  user_lat: 41.6875
                  user_lng: 44.8084
                  status: "new"
                  resolved_at: null
                  created_at: "2025-03-01T14:30:00Z"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOrValidationError'
              example:
                error: "validation_error"
                details:
                  - field: "comment"
                    message: "must be at least 10 characters"
                  - field: "type"
                    message: "must be one of: wrong_location, wrong_fact, inappropriate_content"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /purchases/verify:
    post:
      tags:
      - Purchases
      summary: Verify an in-app purchase
      operationId: verifyPurchase
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPurchaseRequest'
            example:
              platform: "ios"
              transaction_id: "txn_abc123"
              receipt: "MIIbxgYJKoZIhvcNAQcCoIIb..."
              type: "city_pack"
              city_id: 1
              price: 4.99
      responses:
        '201':
          description: Purchase verified and recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
              example:
                data:
                  id: 1
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  type: "city_pack"
                  city_id: 1
                  platform: "ios"
                  transaction_id: "txn_abc123"
                  price: 4.99
                  is_ltd: false
                  expires_at: null
                  created_at: "2025-03-01T15:00:00Z"
        '400':
          description: Invalid receipt or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOrValidationError'
              example:
                error: "validation_error"
                details:
                  - field: "receipt"
                    message: "must not be empty"
                  - field: "price"
                    message: "must be greater than 0"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "unauthorized"
                trace_id: "req-abc-123"
        '409':
          description: Transaction already processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "transaction already processed"
                trace_id: "req-abc-123"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...